requirements.txt
fastapi
uvicorn
opencv-python-headless
pillow
